news_agent/requirements.txt
google-adk
fastapi
uvicorn[standard]